Add `minimal api`

+++ src/server/ContactSystem.Core/Api/Queries/Interfaces/IPaginationQuery.cs
@@ -2,7 +2,7 @@ namespace ContactSystem.Core.Api.Queries.Interfaces;
 
 public interface IPaginationQuery
 {
-	int? Offset { get; }
+	long? Offset { get; }
 
-	int? Limit { get; }
+	long? Limit { get; }
 }
