+++ src/server/ContactSystem.Core/Domain/Contact/Contact.cs
@@ -4,9 +4,9 @@ using Microsoft.EntityFrameworkCore;
 using System.ComponentModel.DataAnnotations;
 
 [Index ( nameof ( Email ) , IsUnique = true )]
-public sealed class Contact : IAuditableModel<int>
+public sealed class Contact : IAuditableModel<long>
 {
-	public int Id { get; set; }
+	public long Id { get; set; }
 
 	[Required]
 	public string? FirstName { get; set; }
@@ -27,29 +27,29 @@ public sealed class Contact : IAuditableModel<int>
 
 	public string? MiddleInitial { get; set; }
 
-	public int CreatedBy { get; set; }
+	public long CreatedBy { get; set; }
 
 	public DateTime Created { get; set; }
 
-	public int? LastModifiedBy { get; set; }
+	public long? LastModifiedBy { get; set; }
 
 	public DateTime? LastModified { get; set; }
 
 	object IAuditable.CreatedBy
 	{
 		get => CreatedBy;
-		set => CreatedBy = ( int ) value;
+		set => CreatedBy = ( long ) value;
 	}
 
 	object? IAuditable.LastModifiedBy
 	{
 		get => LastModifiedBy;
-		set => LastModifiedBy = ( int ) value!;
+		set => LastModifiedBy = ( long ) value!;
 	}
 
 	object IModel.Id
 	{
 		get => Id;
-		set => Id = ( int ) value;
+		set => Id = ( long ) value;
 	}
 }
